requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
networkx>=3.2.1
python-dotenv>=1.0.0
plotly>=5.18.0
numpy>=1.24.0

